update de la toma de tiempos de allIntersect.

+++ codes_sdsl_cst/time_allContained_byQueriesFile.cpp
@@ -75,7 +75,7 @@ int main(int argc, char const *argv[]){
 	cout << "Consultas a realizar: " << num_queries << endl;
 	cout << "Repeticiones a realizar: " << repeticiones << endl;
 
-	unsigned t0, t1;
+	unsigned long long t0, t1;
 
 	gst.statsReset();
 	cout << "Ejecutando consultas en implementación Naive..." << endl;
